Update name of the files to avoid breaking changes
