+++ react/ImageList.tsx
@@ -1,3 +1,4 @@
+import { ListContextProvider, useListContext } from 'vtex.list-context'
 import { useDevice } from 'vtex.device-detector'
 
 import React from 'react'
@@ -5,18 +6,27 @@ import React from 'react'
 import { IMAGE_LIST_SCHEMA } from './utils/schema'
 import { getImagesAsJSXList } from './utils/imageUtils'
 
-interface ImageListProps {
+interface Props {
   images: Image[]
   height: number
 }
 
-const ImageList = (props: ImageListProps) => {
-  const { images, height = 420 } = props
+const ImageList: StorefrontFunctionComponent<Props> = ({
+  images,
+  height = 420,
+  children,
+}) => {
+  const list = useListContext()?.list ?? []
   const { isMobile } = useDevice()
 
   const imageListContent = getImagesAsJSXList(images, isMobile, height)
+  const newListContextValue = list.concat(imageListContent)
 
-  return <React.Fragment>{imageListContent}</React.Fragment>
+  return (
+    <ListContextProvider list={newListContextValue}>
+      {children}
+    </ListContextProvider>
+  )
 }
 
 ImageList.schema = IMAGE_LIST_SCHEMA
